UpdatedUppgift2

+++ src/inputoutput/Uppgift2.java
@@ -10,10 +10,10 @@ public class Uppgift2 {
 		Scanner scanner = new Scanner(System.in);
 		while (true) {
 			System.out.print("Enter a firstnumber: ");
-			int number1 = scanner.nextInt();
+			double number1 = scanner.nextDouble();
 			System.out.print("Enter a secondnumber: ");
-			int number2 = scanner.nextInt();
-			System.out.println("" + number1 + " + " + number2 + " = " + number1 + number2);
+			double number2 = scanner.nextDouble();
+			System.out.println("" + number1 + " + " + number2 + " = " + (number1 + number2));
 			System.out.println("" + number1 + " - " + number2 + " = " + (number1 - number2));
 			System.out.println("" + number1 + " x " + number2 + " = " + number1 * number2);
 			System.out.println("" + number1 + " / " + number2 + " = " + number1 / number2);
